completed/02_GitHub_트렌딩_분석_에이전트/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
lxml>=5.0.0
